add reset grid size functionality
i have separated the js document for it to keep dom
elements in one file, and event elements in the other.

dom file has now createNewGrid element.
To reset the container, i used innerHTML = "", then
I rewrite the whole grid. Thought this would be the
easiest way to tackle this problem.